# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-Ticaret API",
        "description": "JWT kimlik doğrulama ile E-Ticaret uygulaması için API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/orders": {
            "get": {
                "tags": [
                    "Siparişler"
                ],
                "summary": "Kullanıcı siparişlerini getir",
                "description": "Giriş yapmış kullanıcının tüm siparişlerini listeler",
                "operationId": "955bb77a45ae173a485fadb8befed93f",
                "responses": {
                    "200": {
                        "description": "Siparişler başarıyla getirildi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Yetkisiz erişim - Geçersiz veya eksik JWT token"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Siparişler"
                ],
                "summary": "Yeni sipariş oluştur",
                "description": "Sepetteki ürünlerden yeni bir sipariş oluşturur ve sepeti temizler",
                "operationId": "2123ba58dadb22bd6f2c8081af562230",
                "responses": {
                    "201": {
                        "description": "Sipariş başarıyla oluşturuldu",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Sipariş başarıyla oluşturuldu"
                                        },
                                        "data": {
                                            "properties": {
                                                "order": {
                                                    "type": "object"
                                                },
                                                "order_items": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Yetkisiz erişim - Geçersiz veya eksik JWT token"
                    },
                    "422": {
                        "description": "Sepet boş veya yetersiz stok"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders/{id}": {
            "get": {
                "tags": [
                    "Siparişler"
                ],
                "summary": "Belirli bir siparişi getir",
                "description": "ID'si verilen siparişin detaylarını getirir",
                "operationId": "88d27387b1e409e075c18609d4a0c300",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Sipariş ID'si",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sipariş başarıyla getirildi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Yetkisiz erişim - Geçersiz veya eksik JWT token"
                    },
                    "404": {
                        "description": "Sipariş bulunamadı"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login veya register endpoint'inden alınan JWT token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}